streamlit-dashboard/requirements.txt
Flask==2.0.1
numpy==1.21.5
pandas==1.4.2
Pillow==9.3.0
scikit_learn==1.2.0
streamlit==1.16.0
